working!! Cant cancel at last moment

+++ src/main/java/com/mountblue/app/controller/AppointmentController.java
@@ -57,8 +57,16 @@ public class AppointmentController {
     }
     @GetMapping("/delete/{appointmentId}")
     public String delete(@PathVariable("appointmentId") int appointmentId,  Model model, HttpSession session) {
-       AppointmentTime appointmentTime=appointmentService.findById(appointmentId);
-       appointmentService.deleteMeeting(appointmentTime.getDate(),appointmentTime.getFromTime(),appointmentTime.getToTime(),appointmentTime.getEvent().getId());
+       int cancellationTime=30;
+        AppointmentTime appointmentTime=appointmentService.findById(appointmentId);
+       if(appointmentTime.getDate().isEqual(LocalDate.now())&&
+          appointmentTime.getFromTime().isAfter(LocalTime.now().plusMinutes(cancellationTime/*appointmentTime.getEvent().getCancellationTime()*/-1)) )
+        {
+            appointmentService.deleteMeeting(appointmentTime.getDate(),appointmentTime.getFromTime(),appointmentTime.getToTime(),appointmentTime.getEvent().getId());
+        }
+       System.out.println(appointmentTime.getDate().isEqual(LocalDate.now())&&
+               appointmentTime.getFromTime().isAfter(LocalTime.now().plusMinutes(cancellationTime-1)) );
         return "redirect:/event/appointments/"+appointmentTime.getUser().getId();
+
     }
 }
